refactor: enhance layout and component structure for improved UI consistency
- Updated layout styles in layout.tsx and page.tsx to ensure a cohesive design with flexbox for better responsiveness.
- Refactored the AnalysisDashboard component to utilize new Chips and ExpandableList components for improved entity display.
- Introduced a new Chips component for better handling of entity lists and improved visual presentation.
- Removed the obsolete EntityCards component to streamline the codebase and focus on essential features.
- Enhanced the Footer component to accept a className prop for better customization.
