+++ frontend/src/app/page.tsx
@@ -6,18 +6,18 @@ import { Suspense } from "react";
 
 export default function Home() {
   return (
-    <div className="to-coral-50 relative min-h-screen overflow-hidden bg-gradient-to-br from-blue-50 via-slate-50 dark:from-blue-950 dark:via-black dark:to-slate-950">
+    <div className="to-coral-50 relative flex min-h-[calc(100vh-0px)] flex-col overflow-hidden bg-gradient-to-br from-blue-50 via-slate-50 dark:from-blue-950 dark:via-black dark:to-slate-950">
       {/* Hero Section */}
       <HeroSection />
 
       {/* Main Content */}
-      <main className="relative z-10 container mx-auto -mt-8 px-4 pb-16">
+      <main className="relative z-10 container mx-auto -mt-8 flex-1 px-4 pb-16">
         <Suspense fallback={<LoadingSpinner size="lg" />}>
           <TranscriptAnalyzer />
         </Suspense>
       </main>
 
-      <Footer />
+      <Footer className="mt-auto" />
     </div>
   );
 }
