+++ frontend/src/components/ui/chips.tsx
@@ -1,0 +1,43 @@
+import { useState } from "react";
+
+function Chips({ items }: { items: string[] }) {
+  const MAX = 8;
+  const [showAll, setShowAll] = useState(false);
+  const visible = showAll ? items : items.slice(0, MAX);
+
+  if (items.length === 0)
+    return <span className="text-xs text-white/50">—</span>;
+
+  return (
+    <div>
+      <div className="flex flex-wrap gap-1.5">
+        {visible.map((txt) => (
+          <span
+            key={txt}
+            className="rounded-full border border-white/15 bg-white/5 px-2 py-0.5 text-xs text-white/80"
+          >
+            {txt}
+          </span>
+        ))}
+        {items.length > MAX && !showAll && (
+          <button
+            onClick={() => setShowAll(true)}
+            className="rounded-full border border-white/15 px-2 py-0.5 text-xs text-white/70 hover:bg-white/5"
+          >
+            +{items.length - MAX}
+          </button>
+        )}
+      </div>
+      {showAll && items.length > MAX && (
+        <button
+          onClick={() => setShowAll(false)}
+          className="mt-2 text-xs text-[#00A3E0] hover:underline"
+        >
+          Show less
+        </button>
+      )}
+    </div>
+  );
+}
+
+export { Chips };
